fix: reduce gc thrash from cloning guards constantly
