+++ src/Combinators/index.ts
@@ -1,6 +1,5 @@
+export * from './FluentGuardBuilder';
+export * from './altGuard';
+export * from './booleanGuard';
 export * from './notGuard';
 export * from './thenGuard';
-export * from './orGuard';
-export * from './altGuard';
-export * from './andGuard';
-export * from './FluentGuardBuilder';
